refactor: remove subcomponent folder on products component page folder
